Implement deletion strategy for one-child, leaf and two-children nodes
If a delete message is received by the BinaryTree, it switches to a new
context, sets itself as a middleman so that it receives the responses
before the MessageHandler in order to switch back to the original
context and forwards the deletion message to the root node.

From there on, we're just searching for the correct node and if we
finally found it, we send a DeleteChild event to its parent, respond to
the middleman with an OperationFinished event stop the node.

The BinaryTree actor, in the meantime, stores all incoming operations
in a queue, switches back to the original context and works throught the
whole queue.

I decided to use the in-order predecessor, meaning the largest value of
the left subtree as an replacement for the to be deleted node in order
to maintain the ordering of the tree.

A new message MovePredecessor propagates to the node with the "highest"
key, sends DeleteChild to its own parent, replaces its own parent and
its subtree with the ones of the deleted node and sends a ReplaceChild
event to its new parent, in order to be at the place of the old node.